refactor: AutoCompleteUnits
