+++ apps/meteor/client/omnichannel/additionalForms/AutoCompleteUnits.tsx
@@ -3,8 +3,9 @@ import { PaginatedMultiSelectFiltered } from '@rocket.chat/fuselage';
 import { useDebouncedValue } from '@rocket.chat/fuselage-hooks';
 import type { ReactElement } from 'react';
 import { memo, useMemo, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 
-import { useUnits } from '../../components/Omnichannel/hooks/useUnits';
+import { useUnitsList } from '../../components/Omnichannel/hooks/useUnitsList';
 import { useRecordList } from '../../hooks/lists/useRecordList';
 import { AsyncStatePhase } from '../../lib/asyncState';
 
@@ -16,29 +17,32 @@ type AutoCompleteUnitsProps = {
 };
 
 const AutoCompleteUnits = ({ value, error, placeholder, onChange }: AutoCompleteUnitsProps): ReactElement => {
-	const [agentsFilter, setAgentsFilter] = useState<string>('');
+	const { t } = useTranslation();
+	const [unitsFilter, setUnitsFilter] = useState<string>('');
 
-	const debouncedUnitFilter = useDebouncedValue(agentsFilter, 500);
+	const debouncedUnitFilter = useDebouncedValue(unitsFilter, 500);
 
-	const { itemsList: UnitsList, loadMoreItems: loadMoreUnits } = useUnits(
+	const { itemsList: unitsList, loadMoreItems: loadMoreUnits } = useUnitsList(
 		useMemo(() => ({ text: debouncedUnitFilter }), [debouncedUnitFilter]),
 	);
 
-	const { phase: agentsPhase, itemCount: agentsTotal, items: unitItems } = useRecordList(UnitsList);
+	const { phase: unitsPhase, itemCount: unitsTotal, items: unitItems } = useRecordList(unitsList);
 
 	return (
 		<PaginatedMultiSelectFiltered
 			value={value}
 			error={error}
-			placeholder={placeholder}
+			placeholder={placeholder || t('Select_an_option')}
 			onChange={onChange}
+			filter={unitsFilter}
+			width='100%'
 			flexShrink={0}
-			filter={agentsFilter}
-			setFilter={setAgentsFilter as (value: string | number | undefined) => void}
+			flexGrow={0}
+			setFilter={setUnitsFilter as (value: string | number | undefined) => void}
 			options={unitItems}
 			data-qa='autocomplete-multiple-unit'
 			endReached={
-				agentsPhase === AsyncStatePhase.LOADING ? (): void => undefined : (start): void => loadMoreUnits(start!, Math.min(50, agentsTotal))
+				unitsPhase === AsyncStatePhase.LOADING ? (): void => undefined : (start): void => loadMoreUnits(start!, Math.min(50, unitsTotal))
 			}
 		/>
 	);
