+++ apps/meteor/client/views/omnichannel/directory/chats/ChatsFiltersContextualBar.tsx
@@ -17,7 +17,7 @@ import {
 	ContextualbarFooter,
 } from '../../../../components/Contextualbar';
 import { useHasLicenseModule } from '../../../../hooks/useHasLicenseModule';
-import AutoCompleteUnits from '../../../../omnichannel/additionalForms/Units';
+import AutoCompleteUnits from '../../../../omnichannel/additionalForms/AutoCompleteUnits';
 import { CurrentChatTags } from '../../additionalForms';
 import type { ChatsFiltersQuery } from '../contexts/ChatsContext';
 import { useChatsContext } from '../contexts/ChatsContext';
